PO-1941 Remove Defendant Account Party
Swapped if-match with popstedBy

+++ src/main/java/uk/gov/hmcts/opal/dto/legacy/RemoveDefendantAccountPartyLegacyResponse.java
@@ -10,6 +10,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigInteger;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -17,11 +19,10 @@ import lombok.NoArgsConstructor;
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlRootElement(name = "response")
 @XmlType(propOrder = {"version", "defendantAccountParty"})
-
 public class RemoveDefendantAccountPartyLegacyResponse {
 
     @XmlElement(name = "version")
-    private Integer version;
+    private BigInteger version;
 
     @XmlElement(name = "defendant_account_party")
     private DefendantAccountPartyLegacy defendantAccountParty;
